Migrate sessionTab, make sessionfolder stateful
