+++ packages/ui/components/RecipeSidebar/SessionItem.tsx
@@ -1,10 +1,24 @@
 "use client";
-import {
-  RecipeSession,
-  useRecipeSessionStore,
-} from "../../state/recipeSession";
-import { Fragment, useRef, useState } from "react";
+import { useRecipeSessionStore } from "../../state/recipeSession";
+import { useContext, useState } from "react";
 import classNames from "classnames";
+import {
+  Cog6ToothIcon,
+  FolderIcon,
+  FolderOpenIcon,
+  FolderPlusIcon,
+  PlusCircleIcon,
+  ShareIcon,
+} from "@heroicons/react/24/outline";
+import { FolderAPI } from "../../state/apiSession/FolderAPI";
+import { PublishFolderModal } from "../../pages/editor/Builders/PublishModal";
+import { RecipeSessionFolderExtended } from "types/database";
+import { RecipeCloudContext } from "../../state/apiSession/CloudAPI";
+import { FolderModal } from "./Modal/FolderModal";
+import { EditFolderModal } from "./Modal/EditFolderModal";
+
+import { RecipeSession } from "../../state/recipeSession";
+import { Fragment, useRef } from "react";
 import { RouteTypeLabel } from "../RouteTypeLabel";
 import { useHover } from "usehooks-ts";
 import {
@@ -17,12 +31,165 @@ import {
   getConfigForSessionStore,
   getParametersForSessionStore,
 } from "../../state/apiSession";
-import { FolderAPI } from "../../state/apiSession/FolderAPI";
 import { EditSessionModal } from "./Modal/EditSessionModal";
 import { Recipe } from "types/database";
 import { useSupabaseClient } from "../Providers/SupabaseProvider";
 import { cloudEventEmitter } from "../../state/apiSession/CloudAPI";
 import { DuplicateModal } from "./Modal/DuplicateModal";
+
+export function SessionFolder({
+  folder,
+  isRootFolder,
+}: {
+  folder: RecipeSessionFolderExtended;
+  isRootFolder?: boolean;
+}) {
+  const recipeCloud = useContext(RecipeCloudContext);
+  const isCloudFolder =
+    recipeCloud.collectionRecord[folder.id] ??
+    !!recipeCloud.folderToCollection[folder.id];
+
+  const [editFolder, setEditFolder] =
+    useState<RecipeSessionFolderExtended | null>(null);
+  const [publishFolder, setPublishFolder] =
+    useState<RecipeSessionFolderExtended | null>(null);
+  const addEditorSession = useRecipeSessionStore(
+    (state) => state.addEditorSession
+  );
+  const [addFolderModal, setAddFolderModal] = useState(false);
+
+  const [open, setOpen] = useState(true);
+
+  const Icon = open ? FolderOpenIcon : FolderIcon;
+
+  return (
+    <>
+      <li className="w-full">
+        <span
+          className={classNames(
+            " relative w-full text-xs font-bold p-0 px-2 py-2 pr-4 group",
+            open && "menu-dropdown-show"
+          )}
+          onClick={(e) => {
+            e.stopPropagation();
+
+            setOpen(!open);
+          }}
+        >
+          <span
+            className={classNames(
+              "flex items-center w-full",
+              isCloudFolder && "text-blue-500 dark:text-accent"
+            )}
+          >
+            {isCloudFolder ? (
+              <Icon
+                className={classNames(
+                  "w-4 h-4 mr-2 mb-0.5",
+                  isCloudFolder && "text-blue-500 dark:text-accent"
+                )}
+              />
+            ) : (
+              <Icon className="w-4 h-4 mr-2 mb-0.5" />
+            )}
+
+            {folder.name}
+          </span>
+          <div className="hidden absolute right-0 group-hover:flex z-10 bg-base-100 top-0 h-8 px-2 items-center group-hover:border rounded-md">
+            <a
+              className="hidden group-hover:block hover:animate-spin w-fit py-2 px-1"
+              onClick={(e) => {
+                e.preventDefault();
+                e.stopPropagation();
+
+                setEditFolder(folder);
+              }}
+            >
+              <Cog6ToothIcon className="w-4 h-4" />
+            </a>
+            <a
+              className="hidden group-hover:block hover:bg-accent w-fit py-2 px-1"
+              onClick={(e) => {
+                e.preventDefault();
+                e.stopPropagation();
+
+                const session = addEditorSession();
+                FolderAPI.addSessionToFolder(session.id, folder.id);
+              }}
+            >
+              <PlusCircleIcon className="w-4 h-4" />
+            </a>
+            <a
+              className="hidden group-hover:block hover:bg-accent w-fit py-2 px-1"
+              onClick={(e) => {
+                e.preventDefault();
+                e.stopPropagation();
+
+                setAddFolderModal(true);
+              }}
+            >
+              <FolderPlusIcon className="w-4 h-4" />
+            </a>
+            {isRootFolder && (
+              <a
+                className="hidden group-hover:block hover:bg-accent w-fit py-2 px-1"
+                onClick={(e) => {
+                  e.preventDefault();
+                  e.stopPropagation();
+
+                  setPublishFolder(folder);
+                }}
+              >
+                <ShareIcon className="w-4 h-4" />
+              </a>
+            )}
+          </div>
+        </span>
+        {open && (
+          <ul className="">
+            {folder.items.map((item) => {
+              if (!item) return null;
+              if (item.type === "session") {
+                const session = item.session;
+
+                return (
+                  <SessionTab
+                    key={session.id}
+                    session={session}
+                    cloudSession={recipeCloud.apiRecord[session.id]}
+                    parentFolderId={folder.id}
+                  />
+                );
+              } else {
+                const folder = item.folder;
+
+                return <SessionFolder key={folder.id} folder={folder} />;
+              }
+            })}
+          </ul>
+        )}
+      </li>
+      {addFolderModal && (
+        <FolderModal
+          onClose={() => setAddFolderModal(false)}
+          addToFolder={folder}
+        />
+      )}
+      {editFolder && (
+        <EditFolderModal
+          onClose={() => setEditFolder(null)}
+          folder={editFolder}
+        />
+      )}
+      {publishFolder && (
+        <PublishFolderModal
+          onClose={() => setPublishFolder(null)}
+          folder={publishFolder}
+        />
+      )}
+    </>
+  );
+}
 
 export function SessionTab({
   session,
